python_chat_listener/requirements.txt
Flask
Flask-SocketIO
python-dotenv
kafka-python
requests
aiohttp
twitchio